fix: Ensured that all filters are finished
Ensured that all filtering is done during the sinister boot cycle, this
was not done before, but was not causing issues.

Changed the way that filters generate tasks to keep number of concurrent
tasks low, preventing stack overflows unless there are an insane number
of filters.

Also allowed customisation of the filter sync lock.